+++ src/components/swap/CreateSwap.tsx
@@ -117,14 +117,11 @@ export const CreateSwap = () => {
   ]);
 
   const buttonDisabled = useMemo(() => {
-    return !!error.liquidityError ||
-      !!error.insufficientBalanceError ||
-      loadingDisabled ||
-      isSwapping
+    return !!error.liquidityError || loadingDisabled
       ? true
       : needsWalletConnection
         ? false
-        : !isChainSupported
+        : !isChainSupported || isSwapping
           ? true
           : validSwap
             ? false
@@ -154,6 +151,7 @@ export const CreateSwap = () => {
     return getTimeEstimates(inputAsset);
   }, [inputAsset, outputAsset]);
 
+
   const handleConnectWallet = () => {
     if (!needsWalletConnection) return;
 
@@ -176,7 +174,9 @@ export const CreateSwap = () => {
       await fetchAndSetFiatValues();
       await Promise.allSettled([
         address && fetchAndSetEvmBalances(address, workingRPCs),
-        btcAddress && provider && fetchAndSetBitcoinBalance(provider),
+        btcAddress &&
+          provider &&
+          fetchAndSetBitcoinBalance(provider, btcAddress),
         starknetAddress && fetchAndSetStarknetBalance(starknetAddress),
         solanaAnchorProvider &&
           fetchAndSetSolanaBalance(solanaAnchorProvider.publicKey),
@@ -188,8 +188,8 @@ export const CreateSwap = () => {
       if (!inputAsset) return;
       if (isEVM(inputAsset.chain) && address)
         await fetchAndSetEvmBalances(address, workingRPCs, inputAsset);
-      if (isBitcoin(inputAsset.chain) && provider)
-        await fetchAndSetBitcoinBalance(provider);
+      if (isBitcoin(inputAsset.chain) && provider && btcAddress)
+        await fetchAndSetBitcoinBalance(provider, btcAddress);
       if (isStarknet(inputAsset.chain) && starknetAddress)
         await fetchAndSetStarknetBalance(starknetAddress);
       if (isSolana(inputAsset.chain) && solanaAnchorProvider)