+++ src/store/assetInfoStore.ts
@@ -7,11 +7,17 @@ import {
   isBitcoin,
   isEVM,
   isSolana,
+  isSolanaNativeToken,
   isStarknet,
 } from "@gardenfi/orderbook";
 import { API } from "../constants/api";
 import axios from "axios";
-import { Quote, Strategies } from "@gardenfi/core";
+import {
+  BitcoinNetwork,
+  BitcoinProvider,
+  Quote,
+  Strategies,
+} from "@gardenfi/core";
 import { generateTokenKey } from "../utils/generateTokenKey";
 import { PublicKey } from "@solana/web3.js";
 
@@ -27,6 +33,7 @@ import {
   getStarknetTokenBalance,
 } from "../utils/getTokenBalance";
 import { Hex } from "viem";
+import { getSpendableBalance } from "../utils/getmaxBtc";
 
 export type Networks = {
   [chain in Chain]: ChainData & { assetConfig: Omit<AssetConfig, "chain">[] };
@@ -84,7 +91,8 @@ type AssetInfoState = {
     fetchOnlyAsset?: Asset
   ) => Promise<void>;
   fetchAndSetBitcoinBalance: (
-    provider: IInjectedBitcoinProvider
+    provider: IInjectedBitcoinProvider,
+    address: string
   ) => Promise<void>;
   fetchAndSetStarknetBalance: (address: string) => Promise<void>;
   fetchAndSetSolanaBalance: (address: PublicKey) => Promise<void>;
@@ -275,8 +283,12 @@ export const assetInfoStore = create<AssetInfoState>((set, get) => ({
     }
   },
 
-  fetchAndSetBitcoinBalance: async (provider: IInjectedBitcoinProvider) => {
-    const { assets } = get();
+  fetchAndSetBitcoinBalance: async (
+    provider: IInjectedBitcoinProvider,
+
+    address: string
+  ) => {
+    const { assets, balances } = get();
     if (!assets || !provider) return;
 
     try {
@@ -284,25 +296,41 @@ export const assetInfoStore = create<AssetInfoState>((set, get) => ({
       if (!balance?.val?.total) return;
 
       const formattedBalance = new BigNumber(balance.val.confirmed);
+
+      const _provider = new BitcoinProvider(
+        network === "mainnet" ? BitcoinNetwork.Mainnet : BitcoinNetwork.Testnet
+      );
+
+      const feeRate = await _provider.getFeeRates();
+      const utxos = await _provider.getUTXOs(address, Number(formattedBalance));
+      const spendable = await getSpendableBalance(
+        address,
+        Number(formattedBalance),
+        utxos.length,
+        feeRate.fastestFee
+      );
+      const maxSpendableBalance = spendable.ok ? spendable.val : 0;
+
       const btcBalance = Object.values(assets)
         .filter((asset) => isBitcoin(asset.chain))
         .reduce(
           (acc, asset) => {
-            acc[getOrderPair(asset.chain, asset.tokenAddress)] =
-              formattedBalance;
+            acc[getOrderPair(asset.chain, asset.tokenAddress)] = new BigNumber(
+              maxSpendableBalance
+            );
             return acc;
           },
           {} as Record<string, BigNumber | undefined>
         );
 
-      set({ balances: { ...get().balances, ...btcBalance } });
+      set({ balances: { ...balances, ...btcBalance } });
     } catch {
       /*empty*/
     }
   },
 
   fetchAndSetStarknetBalance: async (address: string) => {
-    const { assets } = get();
+    const { assets, balances } = get();
     if (!assets) return;
 
     const starknetAsset = Object.values(assets).find((asset) =>
@@ -319,31 +347,31 @@ export const assetInfoStore = create<AssetInfoState>((set, get) => ({
       starknetAsset.tokenAddress
     );
     starknetBalance[orderPair] = new BigNumber(balance);
-    set({ balances: { ...get().balances, ...starknetBalance } });
+    set({ balances: { ...balances, ...starknetBalance } });
   },
 
   fetchAndSetSolanaBalance: async (address: PublicKey) => {
-    const { assets } = get();
+    const { assets, balances } = get();
     if (!assets) return;
-    const solanaAssets = Object.values(assets).filter((asset) =>
+
+    const solanaAsset = Object.values(assets).find((asset) =>
       isSolana(asset.chain)
     );
 
-    if (solanaAssets.length === 0) return;
+    if (!solanaAsset) return;
+    const solanaBalance: Record<string, BigNumber | undefined> = {};
 
-    const solanaBalances: Record<string, BigNumber | undefined> = {};
+    const balance = await getSolanaTokenBalance(address, solanaAsset);
+    const orderPair = getOrderPair(solanaAsset.chain, solanaAsset.tokenAddress);
 
-    for (const asset of solanaAssets) {
-      try {
-        const balance = await getSolanaTokenBalance(address, asset);
-        const orderPair = getOrderPair(asset.chain, asset.tokenAddress);
-        solanaBalances[orderPair] = new BigNumber(balance);
-      } catch (err) {
-        continue;
-      }
-    }
+    if (isSolanaNativeToken(solanaAsset.chain, solanaAsset.tokenAddress)) {
+      const gas = 0.00380608;
+      solanaBalance[orderPair] = new BigNumber(
+        Math.max(0, Number((Number(balance) - gas).toFixed(8)))
+      );
+    } else solanaBalance[orderPair] = new BigNumber(balance);
 
-    set({ balances: { ...get().balances, ...solanaBalances } });
+    set({ balances: { ...balances, ...solanaBalance } });
   },
 
   clearBalances: () =>